# Conflicts:
#	beestation.dme
#	code/game/objects/effects/glowshroom.dm
#	code/modules/antagonists/blob/blobstrains/cryogenic_poison.dm
#	code/modules/antagonists/blob/blobstrains/networked_fibers.dm
#	code/modules/antagonists/blob/blobstrains/replicating_foam.dm
#	code/modules/mining/lavaland/necropolis_chests.dm
#	code/modules/mob/living/carbon/alien/humanoid/alien_powers.dm
#	code/modules/reagents/chemistry/holder.dm
#	code/modules/reagents/chemistry/machinery/reagentgrinder.dm
#	code/modules/reagents/chemistry/reagents.dm
#	code/modules/reagents/chemistry/reagents/food_reagents.dm
#	code/modules/reagents/chemistry/reagents/medicine_reagents.dm
#	code/modules/reagents/chemistry/reagents/other_reagents.dm
#	code/modules/reagents/chemistry/reagents/pyrotechnic_reagents.dm
#	code/modules/reagents/reagent_containers.dm
#	code/modules/reagents/reagent_containers/cups/drinkingglass.dm
#	code/modules/reagents/reagent_containers/syringes.dm

+++ code/modules/mob/living/simple_animal/friendly/farm_animals.dm
@@ -41,14 +41,14 @@
 	AddComponent(/datum/component/udder)
 	. = ..()
 
-/mob/living/simple_animal/hostile/retaliate/goat/Life()
+/mob/living/simple_animal/hostile/retaliate/goat/Life(delta_time = SSMOBS_DT, times_fired)
 	. = ..()
 	if(.)
 		//chance to go crazy and start wacking stuff
-		if(!enemies.len && prob(1))
+		if(!enemies.len && DT_PROB(0.5, delta_time))
 			Retaliate()
 
-		if(enemies.len && prob(10))
+		if(enemies.len && DT_PROB(5, delta_time))
 			clear_enemies()
 			LoseTarget()
 			src.visible_message(span_notice("[src] calms down."))
@@ -98,78 +98,10 @@
 			H.visible_message(span_warning("[src] takes a big chomp out of [H]!"), \
 									span_userdanger("[src] takes a big chomp out of your [NB]!"))
 			NB.dismember()
+
 /mob/living/simple_animal/hostile/retaliate/goat/rabid
 	name = "Rabid Maintenance Pete"
 	faction = list(FACTION_HOSTILE)
-
-//cow
-/mob/living/simple_animal/cow
-	name = "cow"
-	desc = "Known for their milk, just don't tip them over."
-	icon_state = "cow"
-	icon_living = "cow"
-	icon_dead = "cow_dead"
-	icon_gib = "cow_gib"
-	gender = FEMALE
-	mob_biotypes = list(MOB_ORGANIC, MOB_BEAST)
-	speak = list("moo?","moo","MOOOOOO")
-	speak_emote = list("moos","moos hauntingly")
-	speak_language = /datum/language/metalanguage
-	emote_hear = list("brays.")
-	emote_see = list("shakes its head.")
-	speak_chance = 1
-	turns_per_move = 5
-	see_in_dark = 6
-	butcher_results = list(/obj/item/food/meat/slab = 6)
-	response_help_continuous = "pets"
-	response_help_simple = "pet"
-	response_disarm_continuous = "gently pushes aside"
-	response_disarm_simple = "gently push aside"
-	response_harm_continuous = "kicks"
-	response_harm_simple = "kick"
-	attack_verb_continuous = "kicks"
-	attack_verb_simple = "kick"
-	attack_sound = 'sound/weapons/punch1.ogg'
-	health = 50
-	maxHealth = 50
-	gold_core_spawnable = FRIENDLY_SPAWN
-	blood_volume = BLOOD_VOLUME_NORMAL
-	chat_color = "#FFFFFF"
-
-	footstep_type = FOOTSTEP_MOB_SHOE
-
-/mob/living/simple_animal/cow/Initialize(mapload)
-	AddComponent(/datum/component/udder)
-	. = ..()
-
-/mob/living/simple_animal/cow/attack_hand(mob/living/carbon/M)
-	if(!stat && M.a_intent == INTENT_DISARM && icon_state != icon_dead)
-		M.visible_message(span_warning("[M] tips over [src]."),
-			span_notice("You tip over [src]."))
-		to_chat(src, span_userdanger("You are tipped over by [M]!"))
-		Paralyze(60, ignore_canstun = TRUE)
-		icon_state = icon_dead
-		addtimer(CALLBACK(src, PROC_REF(tip_back), M), rand(20,50))
-	else
-		..()
-
-/mob/living/simple_animal/cow/proc/tip_back(mob/living/carbon/M)
-	if(stat && M)
-		return
-	icon_state = icon_living
-	var/external
-	var/internal
-	switch(pick(1,2,3,4))
-		if(1,2,3)
-			var/text = pick("imploringly.", "pleadingly.",
-				"with a resigned expression.")
-			external = "[src] looks at [M] [text]"
-			internal = "You look at [M] [text]"
-		if(4)
-			external = "[src] seems resigned to its fate."
-			internal = "You resign yourself to your fate."
-	visible_message(span_notice("[external]"),
-		span_revennotice("[internal]"))
 
 /mob/living/simple_animal/chick
 	name = "\improper chick"
@@ -203,6 +135,7 @@
 	health = 3
 	maxHealth = 3
 	ventcrawler = VENTCRAWLER_ALWAYS
+	var/can_grow = TRUE
 	var/amount_grown = 0
 	pass_flags = PASSTABLE | PASSMOB
 	mob_size = MOB_SIZE_TINY
@@ -217,12 +150,14 @@
 	pixel_y = base_pixel_y + rand(0, 10)
 	GLOB.total_chickens++
 
-/mob/living/simple_animal/chick/Life()
+/mob/living/simple_animal/chick/Life(delta_time = SSMOBS_DT, times_fired)
 	. =..()
 	if(!.)
 		return
+	if(!can_grow)
+		return
 	if(!stat && !ckey)
-		amount_grown += rand(1,2)
+		amount_grown += rand(0.5 * delta_time, 1 * delta_time)
 		if(amount_grown >= 100)
 			new /mob/living/simple_animal/chicken(src.loc)
 			qdel(src)
@@ -236,9 +171,13 @@
 		GLOB.total_chickens--
 	return ..()
 
-/mob/living/simple_animal/chick/holo/Life()
-	..()
-	amount_grown = 0
+/mob/living/simple_animal/chick/holo
+	can_grow = FALSE
+
+/mob/living/simple_animal/chick/dave
+	name = "Dave"
+	desc = "A tiny chick personally rescued from the station's kitchen by the captain, now it's the bridge mascot. Despite his previous high density broiler diet, he seemingly does not grow or age."
+	can_grow = FALSE
 
 /mob/living/simple_animal/chicken
 	name = "\improper chicken"
@@ -322,11 +261,11 @@
 	else
 		..()
 
-/mob/living/simple_animal/chicken/Life()
+/mob/living/simple_animal/chicken/Life(delta_time = SSMOBS_DT, times_fired)
 	. =..()
 	if(!.)
 		return
-	if((!stat && prob(3) && eggsleft > 0) && egg_type && GLOB.total_chickens < CONFIG_GET(number/max_chickens))
+	if((!stat && DT_PROB(1.5, delta_time) && eggsleft > 0) && egg_type && GLOB.total_chickens < CONFIG_GET(number/max_chickens))
 		visible_message("[src] [pick(layMessage)]")
 		eggsleft--
 		var/obj/item/E = new egg_type(get_turf(src))
